Use sb-nibbles contrib module, if available
If it is available, the corresponding package is present in the image
when compiling, so use that to conditionally compile legacy support or
support based on sb-nibbles as applicable.  The support based on the
new module simply transforms each nibbles vector accessor into the
corresponding sb-nibbles one.

Tested with SBCL with sb-nibbles, SBCL without sb-nibbles, and Clozure
Common Lisp (all x86-64).