Add portable IEEE floating point support.
New functions for converting between bits and IEEE floating point numbers:
MAKE-SINGLE-FLOAT, SINGLE-FLOAT-BITS, MAKE-DOUBLE-FLOAT, DOUBLE-FLOAT-BITS.
The functions are only used when implementation-specific conversion functions
are not available, since they are slow and do not handle infinities and NaNs.

The new functions have been tested with ECL and CLISP on little endian Linux
and allow those Lisps to pass the tests of https://github.com/brown/protobuf

The code is not conditionally included based on the presence of
:IEEE-FLOATING-POINT in *FEATURES*, since CLISP does not have that feature due
to its lack of support for denormalized numbers.  Partial CLISP support seems
better than nothing.